Relax bind and connect to accept url! as endpoint (experimental)
This relies on url!s being stored and passed to extensions in a way
similar to string!s, which is, strictly speaking, unspecified and
undocumented in the current R3 extension API and could therefore be
subject to change.

Given that we already access string! data directly (via RL_SERIES +
RXI_SER_DATA) instead of using more well-defined R3 Ext API functions
(RL_GET_STRING) this change really isn't half as scary as this comment
may make it sound.

The simple and completely safe approach is, of course, to have
minimalistic bind/connect wrappers written in REBOL which convert url!s
to string!s.